move CFuture/CPromise
